fix broken test after creating with a custom package
Fix (hopefully / untestet) the broken test if you create a project with a custom package. 

For example: if you create a groupId with com.vaadin.addons and later define the package as "awesomeaddon". The final package with your classes is below com.vaadin.addons.awesomeaddon. This won't get recognized here only with the groupId and so the browsertest fails because it can't find the correct js methods.